% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rm.R
\name{plot_rm_score}
\alias{plot_rm_score}
\title{Feature plot of RM prediction scores on the projected UMAP}
\usage{
plot_rm_score(
  query,
  annotation_level = c("H1", "H2", "H3"),
  reduction = "umap.rpca"
)
}
\arguments{
\item{query}{Seurat object after \code{\link[=map_query]{map_query()}}.}

\item{annotation_level}{\code{"H1"}, \code{"H2"}, or \code{"H3"}.}

\item{reduction}{Dimensional reduction to plot. Default \code{"umap.rpca"}.}
}
\value{
A ggplot object (Seurat FeaturePlot).
}
\description{
Feature plot of RM prediction scores on the projected UMAP
}
