% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rm.R
\name{plot_rm_score_hist}
\alias{plot_rm_score_hist}
\title{Histogram of RM prediction scores}
\usage{
plot_rm_score_hist(query, annotation_level = c("H1", "H2", "H3"), breaks = 50)
}
\arguments{
\item{query}{Seurat object after \code{\link[=map_query]{map_query()}}.}

\item{annotation_level}{\code{"H1"}, \code{"H2"}, or \code{"H3"}.}

\item{breaks}{Number of histogram bins. Default \code{50}.}
}
\value{
A ggplot object.
}
\description{
Histogram of RM prediction scores
}
